Adds routing initial setup (using vue-router)

+++ src/components/Body.vue
@@ -1,8 +1,9 @@
 <script setup lang="ts">
 import { ref, onMounted } from "vue";
+
+import type { Todo } from "@/types/todo";
 import BodyItem from "./BodyItem.vue";
 import TodoContainer from "./TodoPill.vue";
-import type { Todo } from "@/types/todo";
 
 const API_BASE_URL = "http://localhost:3000/";
 
@@ -20,14 +21,30 @@ function deleteTodo(todo: Todo) {
 
   if (indexToDelete !== -1) {
     todos.value.splice(indexToDelete, 1);
+
+    // remove the todo from the DB
+    deleteTodoFromDb(indexToDelete);
   }
 }
 
 async function deleteTodoFromDb(idToDelete: number) {
-  const endpoint = API_BASE_URL + "todos";
+  if (!idToDelete) {
+    return;
+  }
+
+  // DELETE from api using id inside the url
+  const endpoint = `${API_BASE_URL}todos/${idToDelete}`;
   const body: RequestInit = {
     method: "DELETE",
   };
+
+  const response = await fetch(endpoint, body);
+  const responseBody = await response.json();
+
+  if (response.ok || response.status === 200) {
+    console.log(responseBody);
+    // todo .. show something to the user
+  }
 }
 
 function addTodo() {
@@ -77,16 +94,23 @@ function synchronizeIds(lastId: number) {
   id = lastId++;
 }
 
-// Lifecycle hooks, you can use them to execute special functions when the component start
-
+// Lifecycle hooks, you can use them to execute special functions on useful moments like component creation ( onMounted )
 onMounted(() => {
   getTodos();
 });
 </script>
 
 <style scoped>
+.todos-wrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-start;
+  margin-left: -30px;
+}
 
-.todo-wrapper{
+.todo-wrapper {
   width: 40rem;
   padding: 15px;
   margin-top: 15px;
@@ -107,34 +131,57 @@ button {
   margin-top: 15px;
 }
 
-.inputs {
+.user-inputs {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
+}
+.action-buttons {
+  margin-top: 30px !important;
+  display: flex;
+  justify-content: flex-start;
   align-items: flex-start;
+}
+
+.margin-right-15 {
+  margin-left: 15px !important;
 }
 </style>
 
 <template>
-  <!-- here we are using <slot/ > to insert an icon -->
   <BodyItem>
     <template #heading>
       <h1>My todo list using vue.js</h1>
+      <h4>
+        You can see your statistics
+
+        <router-link to="statistics">
+          <a href="">here</a>
+        </router-link>
+      </h4>
     </template>
 
-    <div class="todo-wrapper" v-if="todos.length !== 0" v-for="todo in todos" :key="todo.id">
-      <TodoContainer :todo="todo" @on-todo-deletion="deleteTodo(todo)" />
+    <div class="todos-wrapper">
+      <div
+        class="todo-wrapper"
+        v-if="todos.length !== 0"
+        v-for="todo in todos"
+        :key="todo.id"
+      >
+        <TodoContainer :todo="todo" @on-todo-deletion="deleteTodo(todo)" />
+      </div>
+      <p v-else>Loading . . .</p>
     </div>
-    <p v-else>Loading . . .</p>
   </BodyItem>
 
-  <BodyItem>
-    <template #heading>¡Add your todo!</template>
-
-    <div class="inputs">
-      <input type="text" v-model="newTodo" />
+  <div class="user-inputs">
+    <h3>¡Add your todo!</h3>
+    <input type="text" v-model="newTodo" />
+    <div class="action-buttons">
       <button @click="addTodo()">Add</button>
-      <button @click="saveTodos()">Save your todos!</button>
+      <button @click="saveTodos()" class="margin-right-15">
+        Save your todos!
+      </button>
     </div>
-  </BodyItem>
+  </div>
 </template>
